backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.4
supabase==2.11.0
boto3==1.35.81
openai==1.59.4
google-genai==1.5.0
speechmatics-batch==0.4.7
deepgram-sdk==3.7.0
tenacity==9.0.0
python-dotenv==1.0.1
